// enums/slider.ts
export enum SliderTransition {
  Slide = "slide",
  Fade = "fade",
}

export enum SlideChangeDirection {
  Forward = "forward",
  Backward = "backward",
}

export enum BackgroundColor {
  White = "white",
  Black = "black",
  Grey = "grey",
  Red = "red",
  Amber = "amber",
  Green = "green",
  Blue = "blue",
}